# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go-Playground - Random Go/Gin-Boilerplate Playground",
        "title": "Go-Playground",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/redemptions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Create a new redemption request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redemptions"
                ],
                "summary": "Create redemption",
                "parameters": [
                    {
                        "description": "Redemption details",
                        "name": "redemption",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Redemption"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Redemption"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/redemptions/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Get all redemptions for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redemptions"
                ],
                "summary": "Get user redemptions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Redemption"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/redemptions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Get redemption details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redemptions"
                ],
                "summary": "Get redemption by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Redemption ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Redemption"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/redemptions/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Update the status of a redemption",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redemptions"
                ],
                "summary": "Update redemption status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Redemption ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "completed",
                                "pending",
                                "failed",
                                "canceled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Redemption": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant_customers_id": {
                    "type": "string"
                },
                "points_used": {
                    "type": "integer"
                },
                "redemption_date": {
                    "type": "string"
                },
                "reward_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.RedemptionStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.RedemptionStatus": {
            "type": "string",
            "enum": [
                "completed",
                "pending",
                "failed"
            ],
            "x-enum-varnames": [
                "RedemptionStatusCompleted",
                "RedemptionStatusPending",
                "RedemptionStatusFailed"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserIdAuth": {
            "description": "User ID for authentication",
            "type": "apiKey",
            "name": "X-User-Id",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        },
        {
            "UserIdAuth": []
        }
    ]
}